+++ client/src/components/navbar/Navbar.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import Logo from "../../assets/images/logo.svg";
 import "../../styles/navbar.css";
-import MenuNavbar from "./MenuNavbar";
+import MenuBurger from "./MenuBurger";
 
 export default function Navbar() {
   return (
@@ -12,7 +12,7 @@ export default function Navbar() {
           <Link to="/recettes">Recettes</Link>
         </section>
         <img src={Logo} alt="Miam" />
-        <MenuNavbar />
+        <MenuBurger />
         <section>
           <Link to="/inscription">Inscription</Link>
           <button type="button">Connexion</button>
